requirements.txt
django
djangorestframework
mysqlclient
ipdb